add cobra cmd parser
